chore(split rovingtabindex into files): - change related path
