+++ packages/ui/src/components/RovingIndex/index.ts
@@ -1,0 +1,3 @@
+export * from "./RovingItem"
+export * from "./useRovingTabIndex"
+export * from "./RovingTabIndexRoot"
