+++ packages/ui/src/components/Tabs/TabList.tsx
@@ -6,7 +6,7 @@ import {
 } from "react"
 import { useTabContext } from "./useTabContext"
 
-import { RovingTabIndexRoot, useRovingTabIndex } from "./useRovingTabIndex"
+import { RovingTabIndexRoot, useRovingTabIndex } from "../RovingIndex"
 import { Slot } from "@radix-ui/react-slot"
 import { composeRefs } from "../../hooks/useComposedRefs"
 import {
@@ -44,7 +44,7 @@ const setIndicatorStyle = (target: HTMLElement) => {
 export const TabList = ({ children, ...props }: TabListProps) => {
   const { selected } = useTabContext("tab")
   return (
-    <RovingTabIndexRoot as="div" active={selected}>
+    <RovingTabIndexRoot active={selected}>
       <div role="tablist" {...props}>
         {children}
       </div>
